requirements.txt

requests>=2.31.0
numpy
pandas
python-dotenv>=1.0.0
yfinance

